more quakeforge compiling on alpha patches

+++ libs/video/renderer/sw/d_edge.c
@@ -172,7 +172,7 @@ D_DrawSurfaces (void)
 			d_zistepv = s->d_zistepv;
 			d_ziorigin = s->d_ziorigin;
 
-			D_DrawSolidSurface (s, ((int) s->data & 0xFF));
+			D_DrawSolidSurface (s, ((size_t) s->data & 0xFF));
 			D_DrawZSpans (s->spans);
 		}
 	} else {
